requirements.txt
python-telegram-bot[job-queue]
pillow
pytesseract
gspread
oauth2client
